Everywhere: Add test-js windows support

+++ AK/FloatingPoint.h
@@ -60,8 +60,8 @@ union FloatExtractor<f64> {
     static constexpr unsigned exponent_max = 2047;
     struct {
         unsigned long long mantissa : 52;
-        unsigned exponent : 11;
-        unsigned sign : 1;
+        u64 exponent : 11;
+        u64 sign : 1;
     };
     f64 d;
 };
